Lustige fix

+++ Backend/src/main/java/com/coderbunker/supqr/service/ObjectService.java
@@ -6,10 +6,14 @@ import java.io.ByteArrayInputStream;
 import java.util.List;
 
 import javax.inject.Inject;
+import javax.ws.rs.ServerErrorException;
+import javax.ws.rs.core.Response.Status;
 
 import org.jooq.generated.tables.pojos.Feedback;
 
 import com.coderbunker.supqr.annotation.Injectable;
+import com.coderbunker.supqr.auth.User;
+import com.coderbunker.supqr.auth.User.UserType;
 import com.coderbunker.supqr.database.FeedbackRepository;
 import com.coderbunker.supqr.database.ObjectRepository;
 import com.coderbunker.supqr.rest.model.CreateObjectTO;
@@ -94,9 +98,10 @@ public class ObjectService {
 			.build();
 	}
 
-	public void deleteObject (Integer objectId) {
-		// TODO: Check if user is allowed to delete
-
+	public void deleteObject (Integer objectId, User user) {
+		if (objectRepository.isUserAuthorOfArticle(objectId, user.getUserId()) && user.getUserType() != UserType.ADMIN) {
+			throw new ServerErrorException(Status.UNAUTHORIZED);
+		}
 		objectRepository.deleteObject(objectId);
 	}
 }
